LayoutTests/accessibility/mac/details-summary-expected.txt

Some open info
Details about the open topic.

Some open info
Some open info
Details about the open topic.

This tests some basic attributes about the details element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS details1.role is 'AXRole: AXGroup'
PASS details1.subrole is 'AXSubrole: AXDetails'
PASS details1.isExpanded is true
PASS summary1.role is 'AXRole: AXButton'
PASS summary1.subrole is 'AXSubrole: AXSummary'
PASS summary1.title is 'AXTitle: Some open info'
PASS details1.isAttributeSettable('AXExpanded') is true
PASS details1.isExpanded is false
PASS summary1.isExpanded is false
PASS details1.isExpanded is false
PASS summary1.isExpanded is false
PASS details1.isExpanded is true
PASS summary1.isExpanded is true
PASS details1.isExpanded is true
PASS summary1.isExpanded is true
PASS details2.subrole is 'AXSubrole: AXDetails'
PASS details2.isExpanded is false
PASS details3.subrole is 'AXSubrole: AXApplicationGroup'
PASS details3.isExpanded is true
Received AXExpandedChanged notification
Received AXExpandedChanged notification
PASS successfullyParsed is true

TEST COMPLETE
